data_governance/requirements.txt
clickhouse-connect
python-dotenv
pandas
oracledb